pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.8",
    "firebase-admin>=6.6.0",
    "openai>=1.63.2",
    "pydantic-settings>=2.8.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
